backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
SQLAlchemy==2.0.43
alembic==1.16.5
